+++ packages/react-router-dom/index.tsx
@@ -74,7 +74,6 @@ export type {
   ActionFunction,
   ActionFunctionArgs,
   AwaitProps,
-  DataMemoryRouterProps,
   DataRouteMatch,
   DataRouteObject,
   Fetcher,