+++ packages/react-router-dom-v5-compat/CHANGELOG.md
@@ -1,8 +1,16 @@
-# react-router-dom-v5-compat
+# `react-router-dom-v5-compat`
+
+## 6.4.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@6.4.1`
+  - `react-router-dom@6.4.1`
 
 ## 6.4.0
 
 **Updated dependencies**
 
-- react-router-dom@6.4.0
-- react-router@6.4.0
+- `react-router-dom@6.4.0`
+- `react-router@6.4.0`