+++ packages/react-router/CHANGELOG.md
@@ -1,4 +1,12 @@
-# react-router
+# `react-router`
+
+## 6.4.1
+
+### Patch Changes
+
+- Preserve state from `initialEntries` ([#9288](https://github.com/remix-run/react-router/pull/9288))
+- Updated dependencies:
+  - `@remix-run/router@1.0.1`
 
 ## 6.4.0
 
@@ -19,7 +27,7 @@ Whoa this is a big one! `6.4.0` brings all the data loading and mutation APIs ov
 
 **Updated Dependencies**
 
-- @remix-run/router@1.0.0
+- `@remix-run/router@1.0.0`
 
 [rr-docs]: https://reactrouter.com/
 [rr-feature-overview]: https://reactrouter.com/en/6.4.0/start/overview