+++ packages/router/CHANGELOG.md
@@ -1,4 +1,11 @@
-# @remix-run/router
+# `@remix-run/router`
+
+## 1.0.1
+
+### Patch Changes
+
+- Preserve state from `initialEntries` ([#9288](https://github.com/remix-run/react-router/pull/9288))
+- Preserve `?index` for fetcher get submissions to index routes ([#9312](https://github.com/remix-run/react-router/pull/9312))
 
 ## 1.0.0
 
@@ -6,7 +13,7 @@ This is the first stable release of `@remix-run/router`, which provides all the 
 
 For an overview of the features provided by `react-router`, we recommend you go check out the [docs][rr-docs], especially the [feature overview][rr-feature-overview] and the [tutorial][rr-tutorial].
 
-For an overview of the features provided by `@remix-run/router`, please check out the [README][remix-router-readme].
+For an overview of the features provided by `@remix-run/router`, please check out the [`README`][remix-router-readme].
 
 [rr-docs]: https://reactrouter.com/
 [rr-feature-overview]: https://reactrouter.com/en/6.4.0/start/overview